+++ Packages/TableProCore/Sources/TableProMSSQLCore/MSSQLConnectionOptions.swift
@@ -7,7 +7,7 @@ public struct MSSQLConnectionOptions: Sendable, Equatable {
     public var password: String
     public var database: String
     public var schema: String
-    public var encryptionFlag: String
+    public var encryptionLevel: MSSQLEncryptionLevel
     public var applicationName: String
     public var loginTimeoutSeconds: Int
     public var authMethod: MSSQLAuthMethod
@@ -27,7 +27,7 @@ public struct MSSQLConnectionOptions: Sendable, Equatable {
     public static let defaultPort = 1_433
     public static let defaultSchema = "dbo"
     public static let defaultApplicationName = "TablePro"
-    public static let defaultEncryptionFlag = "off"
+    public static let defaultEncryptionLevel: MSSQLEncryptionLevel = .request
     public static let defaultLoginTimeoutSeconds = 30
 
     public init(
@@ -37,7 +37,7 @@ public struct MSSQLConnectionOptions: Sendable, Equatable {
         password: String,
         database: String,
         schema: String = MSSQLConnectionOptions.defaultSchema,
-        encryptionFlag: String = MSSQLConnectionOptions.defaultEncryptionFlag,
+        encryptionLevel: MSSQLEncryptionLevel = MSSQLConnectionOptions.defaultEncryptionLevel,
         applicationName: String = MSSQLConnectionOptions.defaultApplicationName,
         loginTimeoutSeconds: Int = MSSQLConnectionOptions.defaultLoginTimeoutSeconds,
         authMethod: MSSQLAuthMethod = .sqlServer,
@@ -59,7 +59,7 @@ public struct MSSQLConnectionOptions: Sendable, Equatable {
         }
         self.database = database
         self.schema = schema
-        self.encryptionFlag = encryptionFlag
+        self.encryptionLevel = encryptionLevel
         self.applicationName = applicationName
         self.loginTimeoutSeconds = loginTimeoutSeconds
     }